examples/ffi/pyproject.toml
[build-system]
requires = ["scikit-build-core", "nanobind", "olympus>=0.4.31"]
build-backend = "scikit_build_core.build"

[project]
name = "olympus_ffi_example"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = ["olympus"]

[project.optional-dependencies]
test = ["pytest", "absl-py"]
